requirements.txt
django==1.11
django-leaflet
psycopg2
djangorestframework
markdown
django-filter
Pillow
pyshp